backend/requirements.txt
fastapi
uvicorn
sqlalchemy
psycopg2-binary
pydantic
pydantic-settings
python-dotenv
python-multipart
jose[cryptography]
passlib[bcrypt]
pandas
numpy
google-genai
fastapi-mail
kiteconnect
langchain
langchain-community
langchain-google-genai
langchain-exa
langgraph
python-pdf
